Improve EventFilter instantiation
Event filters were only able to be created with Ignorefiles.
This meant that the ignore_files configuration value was not
being used whatsoever. Now both an array of files to explicitly
ignore as well as the Ignorefiles can be loaded.

This should also help take care of reducing a bunch of
noise that would happen when initializing git repositories and
doing other gitty things.