Handle when Shopify can't download theme
Github master.zips aren't cached for a long time
so there are times where the `src` isn't actually
available because it hasn't been built yet. If that
is the case, simply retry, it should be good.

Only retry a few times, if we keep failing, simply
report an error and crash