Adds Release script to enable automatic updating
In order to perform automatic updates we need to know
where the new binary is going to be. While it is possible
to include that in the binary itself, that's fragile.

This is a basic script that builds provides two feeds
along with directories containing the new binary for
each platform.

Clients can either pull down changes from latest.json
or if the version difference is too large apply changes
one at a time (more important if we move over to binary
delta patches).

Right now we'll just upload new uncompressed binaries to
S3 each time and add a new record to the all.json feed.

There is also some guard code in the release script to
ensure that we don't attempt to create an update if
the build hasn't been run or if we've already uploaded
that version.